backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-gis==1.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
gunicorn==21.2.0
django-admin-interface==0.26.0
django-colorfield==0.11.0